Typo fixes
- Define filterset -> graphene type generator
- Fix is_test issues
- Move filter_set and test to deep_explore directory
