+++ deep/caches.py
@@ -24,18 +24,18 @@ class CacheKey:
 
     class ExploreDeep:
         _PREFIX = 'EXPLORE-DEEP-{}-'
-        TOTAL_PROJECTS = _PREFIX + 'TOTAL-PROJECTS'
-        TOTAL_REGISTERED_USERS = _PREFIX + 'TOTAL-REGISTERED-USERS'
-        TOTAL_LEADS = _PREFIX + 'TOTAL-LEADS'
-        TOTAL_ENTRIES = _PREFIX + 'TOTAL-ENTRIES'
-        TOTAL_ACTIVE_USERS = _PREFIX + 'TOTAL-ACTIVE-USERS'
-        TOTAL_AUTHORS = _PREFIX + 'TOTAL-AUTHORS'
-        TOTAL_PUBLISHERS = _PREFIX + 'TOTAL-PUBLISHERS'
-        TOP_TEN_AUTHORS = _PREFIX + 'TOP-TEN-AUTHORS'
-        TOP_TEN_PUBLISHERS = _PREFIX + 'TOP-TEN-PUBLISHERS'
-        TOP_TEN_FRAMEWORKS = _PREFIX + 'TOP-TEN-FRAMEWORKS'
-        TOP_TEN_PROJECTS_BY_USERS = _PREFIX + 'TOP-TEN-PROJECTS-BY-USERS'
-        TOP_TEN_PROJECTS_BY_ENTRIES = _PREFIX + 'TOP-TEN-PROJECTS-BY-ENTRIES'
+        TOTAL_PROJECTS_COUNT = _PREFIX + 'TOTAL-PROJECTS'
+        TOTAL_REGISTERED_USERS_COUNT = _PREFIX + 'TOTAL-REGISTERED-USERS'
+        TOTAL_LEADS_COUNT = _PREFIX + 'TOTAL-LEADS'
+        TOTAL_ENTRIES_COUNT = _PREFIX + 'TOTAL-ENTRIES'
+        TOTAL_ACTIVE_USERS_COUNT = _PREFIX + 'TOTAL-ACTIVE-USERS'
+        TOTAL_AUTHORS_COUNT = _PREFIX + 'TOTAL-AUTHORS'
+        TOTAL_PUBLISHERS_COUNT = _PREFIX + 'TOTAL-PUBLISHERS'
+        TOP_TEN_AUTHORS_LIST = _PREFIX + 'TOP-TEN-AUTHORS'
+        TOP_TEN_PUBLISHERS_LIST = _PREFIX + 'TOP-TEN-PUBLISHERS'
+        TOP_TEN_FRAMEWORKS_LIST = _PREFIX + 'TOP-TEN-FRAMEWORKS'
+        TOP_TEN_PROJECTS_BY_USERS_LIST = _PREFIX + 'TOP-TEN-PROJECTS-BY-USERS'
+        TOP_TEN_PROJECTS_BY_ENTRIES_LIST = _PREFIX + 'TOP-TEN-PROJECTS-BY-ENTRIES'
 
 
 class CacheHelper:
