+++ schema.graphql
@@ -1756,6 +1756,7 @@ input LeadsFilterDataInputType {
 }
 
 type LeadsFilterDataType {
+  entriesFilterData: EntriesFilterDataType
   createdBy: [ID!]
   modifiedBy: [ID!]
   ids: [ID!]
@@ -1790,7 +1791,6 @@ type LeadsFilterDataType {
   emmRiskFactors: String
   hasDuplicates: Boolean
   duplicatesOf: ID
-  entriesFilterData: EntriesFilterDataType
 }
 
 type Login {
